Admin\Order\EditControllerTest の抽象クラス作成
- 複数配送用の EditControllerTest を作成したいため
